requirements.txt
flask
msal
azure-storage-blob
azure-identity
pyodbc
routes
services